more bugfixes in auroracam

+++ auroracam/checkAuroracam.sh
@@ -4,12 +4,14 @@ filetocheck=$DATADIR/../live.jpg
 
 while true
 do
-   x=$(find ${filetocheck} -mmin +5)
-   if [ "$x" !=  "" ] ; then
-     echo it crashed
-     sudo reboot
-   else
-     echo all ok
-   fi
-   sleep 30
+  if [ ! -f $DATADIR/../.noreboot ] ; then 
+    x=$(find ${filetocheck} -mmin +5)
+    if [ "$x" !=  "" ] ; then
+      echo it crashed
+      sudo reboot
+    else
+      echo all ok
+    fi
+  fi
+  sleep 30
 done
