+++ auroracam/grabImage.py
@@ -24,8 +24,11 @@ def getStartEndTimes(datadir):
         # we are starting after dusk so find out if there's already a folder
         # and use that instead
         log.info('after overnight start time')
-        starttime = datetime.datetime.now()
+
+        # if starttime=True, then dur is the number of seconds from now to end time.
+        starttime = datetime.datetime.utcnow()
         endtime = starttime + datetime.timedelta(seconds=dur)
+        # see if there's an existing folder for the data
         dirs=[]
         flist = os.listdir(datadir)
         for fl in flist:
@@ -33,11 +36,22 @@ def getStartEndTimes(datadir):
                 dirs.append(fl)
         dirs.sort()
         laststart = datetime.datetime.strptime(dirs[-1], '%Y%m%d_%H%M%S')
+        log.info(f'found folder {dirs[-1]}')
+
+        # if its between noon and midnight on the same day, reuse the folder
         if starttime.hour > 12 and laststart.day == starttime.day:
+            log.info(f'using {dirs[-1]}')
             starttime = laststart
-        elif starttime.hour <= 12 and (starttime.day - laststart.day == 1):
+
+        # if its between midnight and noon, and the dates are less than two full 
+        # days apart, reuse the  folder. 
+        elif starttime.hour <= 12 and (starttime - laststart).days < 2:
+            log.info(f'using {dirs[-1]}')
             starttime = laststart
+        else:
+            pass
     else:
+        # start time is in the future, so add on  dur seconds to get end time
         endtime = starttime + datetime.timedelta(seconds=dur)
     log.info(f'night starts at {starttime} and ends at {endtime}')
     return starttime, endtime
@@ -127,15 +141,18 @@ if __name__ == '__main__':
 
     nightgain = int(os.getenv('NIGHTGAIN', default='70'))
 
-    datadir = os.getenv('DATADIR', default=os.path.expanduser('~/RMS_data/logs'))
+    datadir = os.getenv('DATADIR', default=os.path.expanduser('~/RMS_data/auroracam'))
     os.makedirs(datadir, exist_ok=True)
-
+    norebootflag = os.path.join(datadir, '..', '.noreboot')
+    if os.path.isfile(norebootflag):
+        os.remove(norebootflag)
+    
     # get todays dusk and tomorrows dawn times
     dusk, dawn = getStartEndTimes(datadir)
     dirnam = os.path.join(datadir, dusk.strftime('%Y%m%d_%H%M%S'))
     os.makedirs(dirnam, exist_ok=True)
 
-    now = datetime.datetime.now()
+    now = datetime.datetime.utcnow()
     if now > dawn or now < dusk:
         isnight = False
         setCameraExposure(ipaddress, 'DAY', nightgain, True)
@@ -146,7 +163,7 @@ if __name__ == '__main__':
     log.info(f'now {now}, night start {dusk}, end {dawn}')
     uploadcounter = 0
     while True:
-        now = datetime.datetime.now()
+        now = datetime.datetime.utcnow()
         if now < dawn and now > dusk:
             isnight = True
             setCameraExposure(ipaddress, 'NIGHT', nightgain, True)
@@ -165,6 +182,8 @@ if __name__ == '__main__':
             log.info(f'grabbed {fnam}')
             if isnight is True:
                 # make the mp4
+                norebootflag = os.path.join(datadir, '..', '.noreboot')
+                open(norebootflag, 'w')
                 makeTimelapse(dirnam, s3, 'UK9999')
                 # refresh the dusk/dawn times for tomorrow
                 dusk, dawn = getStartEndTimes(datadir)
@@ -173,6 +192,7 @@ if __name__ == '__main__':
                 setCameraExposure(ipaddress, 'DAY', nightgain, True)
                 log.info('switched to daytime mode, now rebooting')
                 isnight = False
+                os.remove(norebootflag)
                 os.system('/usr/bin/sudo /usr/sbin/shutdown -r now')
 
         # otherwise its night time so save a dated file
@@ -184,8 +204,11 @@ if __name__ == '__main__':
             grabImage(ipaddress, fnam, camid, now)
             log.info(f'grabbed {fnam}')
             fnam2 = os.path.expanduser(os.path.join('~/RMS_data', 'live.jpg'))
-            shutil.copy(fnam, fnam2)
-            log.info('updated live copy')
+            if os.path.isfile(fnam):
+                shutil.copy(fnam, fnam2)
+                log.info('updated live copy')
+            else:
+                log.info(f'{fnam} missing')
 
         uploadcounter += pausetime
         testmode = int(os.getenv('TESTMODE', default=0))
